+++ src/lib/components/composed/json/index.ts
@@ -1,7 +1,4 @@
-import JsonView from "./json-view.svelte";
-import {type JsonViewerProps} from "./json-view.svelte";
+import JsonView from './json-view.svelte';
+import { type JsonViewerProps } from './json-view.svelte';
 
-export { 
-    JsonView,
-    type JsonViewerProps 
-};
+export { JsonView, type JsonViewerProps };