+++ src/utils/assert.ts
@@ -5,7 +5,7 @@ export class AssertionError extends Error {
 	/** @param message The error message. */
 	constructor(message: string) {
 		super(message);
-		this.name = "AssertionError";
+		this.name = 'AssertionError';
 	}
 }
 
@@ -15,7 +15,7 @@ export class AssertionError extends Error {
  * @param expr The expression to test.
  * @param msg The message to display if the assertion fails.
  */
-export function assert(expr: unknown, msg = ""): asserts expr {
+export function assert(expr: unknown, msg = ''): asserts expr {
 	if (!expr) {
 		throw new AssertionError(msg);
 	}