+++ vite.json.config.ts
@@ -4,29 +4,29 @@ import { svelte } from '@sveltejs/vite-plugin-svelte';
 import path from 'path';
 
 export const baseConfig: UserConfig = {
-    resolve: {
-        alias: {
-            $lib: path.resolve("./src/lib").replaceAll("\\", "/"),
-        },
-    },
-    plugins: [tailwindcss(), svelte()],
-    build: {
-        assetsDir: '',
+	resolve: {
+		alias: {
+			$lib: path.resolve('./src/lib').replaceAll('\\', '/')
+		}
+	},
+	plugins: [tailwindcss(), svelte()],
+	build: {
+		assetsDir: '',
 		outDir: './py/corr_vars_widget/static/json',
-        lib: {
-            entry: ["./src/json.ts"],
-            cssFileName: "main",
-            formats: ["es"],
-        },
+		lib: {
+			entry: ['./src/json.ts'],
+			cssFileName: 'main',
+			formats: ['es']
+		},
 		rollupOptions: {
 			output: {
 				entryFileNames: `[name].js`,
 				chunkFileNames: `[name].js`,
-				assetFileNames: `[name].[ext]`,
+				assetFileNames: `[name].[ext]`
 			}
 		}
 	}
-}
+};
 
 // https://vite.dev/config/
 export default defineConfig(baseConfig);